opus: Fix frame size calculation for stereo
There was a bug that frame size was not calculated correctly for stereo
decoding. This requires knowning whether the decoder is for mono or
stereo playback and thus add that to Decoder struct. Rename the variable
in Decoder::decode() accordingly.

Fix alignment of decoder. It seems that there is some requirement for
alignment from libopus and that matters on ARM.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>