opus-embedded-sys: Enable documentation
libopus has documentation in docxygen format. Include it in the build
and convert it to rustdoc. Use doxygen_bindgen for conversion with some
custom replacements on top.

Since there doesn't seem to be a way to add more than one type of
callbacks to bindgen's Builder thus combine CargoCallbacks with our own
comment callback.

Since the output is still not quite well-formatted, include links to
original documentation. That may come handy anyway.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>